Added wait time between actions
Changed check_matches() and move_down() so that they now work on lists of blocks instead of single blocks. Added a wait time in between their actions so it's easier to understand what's happening and why. Added screen updates in both functions to illustrate the wait times better. Also changed the way they're called in add_column() so that they operate on on blocks in the right order.